Reorganize assets/ files
