+++ assets/loading/AssetLump.h
@@ -14,18 +14,10 @@
 #include <filesystem>
 #include <iostream>
 
-#include "assets/AssetTypes.h"
+#include "assets/format/AssetTypes.h"
 
 namespace mondradiko {
 namespace assets {
-
-using LumpHash = uint32_t;
-
-enum class LumpHashMethod : uint8_t { None = 0, xxHash = 1 };
-
-enum class LumpCompressionMethod : uint8_t { None = 0, LZ4 = 1 };
-
-std::string generateLumpName(uint32_t);
 
 class AssetLump {
  public:
